refactor: remove deep nesting of files
complexity not required for this app yet
