+++ package.json
@@ -44,7 +44,7 @@
     "prettier": "^1.11.0"
   },
   "dependencies": {
-    "@blueprintjs/core": "^2.0.0-rc.2",
+    "@blueprintjs/core": "^2.0.0",
     "array-move": "^1.0.0",
     "electron-compile": "^6.4.2",
     "electron-store": "^1.3.0",
