+++ src/components/ActivityIconAndName.js
