--- src/picker/Activity/index.js
@@ -1,1 +1,0 @@
-export { default } from './Activity.container'
