+++ src/picker/components/Activity.container.js
