+++ src/prefs/components/Tab.About.js
@@ -1,14 +1,14 @@
 import { remote, shell } from 'electron'
 import React from 'react'
 
-import UpdateStatus from '../../modules/UpdateStatus'
+import UpdateStatus from './UpdateStatus.container'
 
 const About = () => {
   return (
     <div style={{ textAlign: 'center' }}>
       <p>
         <img
-          src="../../images/icon/icon.png"
+          src="../images/icon/icon.png"
           height="100"
           width="100"
           alt=""
