--- src/prefs/modules/ActivitiesList/index.js
@@ -1,1 +1,0 @@
-export { default } from './ActivitiesList.container'
