--- src/prefs/modules/UpdateStatus/index.js
@@ -1,1 +1,0 @@
-export { default } from './UpdateStatus.container'
