--- src/prefs/tabs/About/index.js
@@ -1,1 +1,0 @@
-export { default } from './About'
