--- src/prefs/tabs/Activities/index.js
@@ -1,1 +1,0 @@
-export { default } from './Activities'
